add image api

+++ modules/clay_core/include/clay_core/blob.h
@@ -7,9 +7,9 @@ namespace clay
 namespace core
 {
 struct IBlob {
-    virtual u8*   get_data() const noexcept = 0;
-    virtual usize get_size() const noexcept = 0;
-    virtual ~IBlob() noexcept               = default;
+    virtual u8* get_data() const noexcept = 0;
+    virtual u64 get_size() const noexcept = 0;
+    virtual ~IBlob() noexcept             = default;
 };
 } // namespace core
 } // namespace clay
